working on user input- mainly talked to trevor and luke
trevor suggested a diff path in terms of complexity as well of complexity. I drew another mock up of my microgame i can send it if you want. Now that i know what im doing for input im gonna make a sprite sheet for the letters. also trevor is not on board for the autism creature in the final project lol but that's not a big deal